Switch on trip at reconnect for this the CameraSettingsForm and uc_CameraSettings moved to GCSView project
