+++ ExtLibs/MV04.Camera/MV04.Camera.csproj
@@ -75,12 +75,6 @@
     <Compile Include="CameraMoverForm.Designer.cs">
       <DependentUpon>CameraMoverForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="CameraSettingsForm.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="CameraSettingsForm.Designer.cs">
-      <DependentUpon>CameraSettingsForm.cs</DependentUpon>
-    </Compile>
     <Compile Include="GCSModeTesterForm.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -108,9 +102,6 @@
     </EmbeddedResource>
     <EmbeddedResource Include="CameraMoverForm.resx">
       <DependentUpon>CameraMoverForm.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="CameraSettingsForm.resx">
-      <DependentUpon>CameraSettingsForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="GCSModeTesterForm.resx">
       <DependentUpon>GCSModeTesterForm.cs</DependentUpon>
