+++ GCSViews/CameraSettingsForm.resx
