+++ GCSViews/CameraView.resx
@@ -328,7 +328,7 @@
     <value>63, 0</value>
   </data>
   <data name="lb_DroneStatusValue.Size" type="System.Drawing.Size, System.Drawing">
-    <value>46, 16</value>
+    <value>37, 13</value>
   </data>
   <data name="lb_DroneStatusValue.TabIndex" type="System.Int32, mscorlib">
     <value>6</value>
@@ -358,7 +358,7 @@
     <value>63, 74</value>
   </data>
   <data name="lb_GCSSelectedStateValue.Size" type="System.Drawing.Size, System.Drawing">
-    <value>62, 16</value>
+    <value>53, 13</value>
   </data>
   <data name="lb_GCSSelectedStateValue.TabIndex" type="System.Int32, mscorlib">
     <value>9</value>
@@ -394,7 +394,7 @@
     <value>0, 0</value>
   </data>
   <data name="lb_GCSSelectedState.Size" type="System.Drawing.Size, System.Drawing">
-    <value>45, 17</value>
+    <value>36, 13</value>
   </data>
   <data name="lb_GCSSelectedState.TabIndex" type="System.Int32, mscorlib">
     <value>8</value>
@@ -454,7 +454,7 @@
     <value>0, 0</value>
   </data>
   <data name="lb_Drone.Size" type="System.Drawing.Size, System.Drawing">
-    <value>57, 17</value>
+    <value>45, 13</value>
   </data>
   <data name="lb_Drone.TabIndex" type="System.Int32, mscorlib">
     <value>4</value>
@@ -514,7 +514,7 @@
     <value>0, 0</value>
   </data>
   <data name="lb_Camera.Size" type="System.Drawing.Size, System.Drawing">
-    <value>68, 17</value>
+    <value>53, 13</value>
   </data>
   <data name="lb_Camera.TabIndex" type="System.Int32, mscorlib">
     <value>5</value>
@@ -568,7 +568,7 @@
     <value>63, 37</value>
   </data>
   <data name="lb_CameraStatusValue.Size" type="System.Drawing.Size, System.Drawing">
-    <value>60, 16</value>
+    <value>49, 13</value>
   </data>
   <data name="lb_CameraStatusValue.TabIndex" type="System.Int32, mscorlib">
     <value>7</value>
@@ -844,7 +844,7 @@
     <value>cs_ColorSliderAltitude</value>
   </data>
   <data name="&gt;&gt;cs_ColorSliderAltitude.Type" xml:space="preserve">
-    <value>MissionPlanner.Controls.ColorSlider, MissionPlanner, Version=1.3.9054.30596, Culture=neutral, PublicKeyToken=null</value>
+    <value>MissionPlanner.Controls.ColorSlider, MissionPlanner, Version=1.3.9055.18093, Culture=neutral, PublicKeyToken=null</value>
   </data>
   <data name="&gt;&gt;cs_ColorSliderAltitude.Parent" xml:space="preserve">
     <value>tlp_AGLContainer</value>
@@ -973,10 +973,10 @@
     <value>NoControl</value>
   </data>
   <data name="lb_AltitudeValue.Location" type="System.Drawing.Point, System.Drawing">
-    <value>13, 37</value>
+    <value>24, 37</value>
   </data>
   <data name="lb_AltitudeValue.Size" type="System.Drawing.Size, System.Drawing">
-    <value>113, 37</value>
+    <value>90, 32</value>
   </data>
   <data name="lb_AltitudeValue.TabIndex" type="System.Int32, mscorlib">
     <value>6</value>
@@ -1009,10 +1009,10 @@
     <value>NoControl</value>
   </data>
   <data name="lb_Agl.Location" type="System.Drawing.Point, System.Drawing">
-    <value>31, 0</value>
+    <value>38, 5</value>
   </data>
   <data name="lb_Agl.Size" type="System.Drawing.Size, System.Drawing">
-    <value>77, 37</value>
+    <value>62, 32</value>
   </data>
   <data name="lb_Agl.TabIndex" type="System.Int32, mscorlib">
     <value>5</value>
@@ -1129,7 +1129,7 @@
     <value>Fill</value>
   </data>
   <data name="btn_TripSwitchOnOff.Font" type="System.Drawing.Font, System.Drawing">
-    <value>Stencil, 20.25pt</value>
+    <value>Stencil, 12pt</value>
   </data>
   <data name="btn_TripSwitchOnOff.ImeMode" type="System.Windows.Forms.ImeMode, System.Windows.Forms">
     <value>NoControl</value>
@@ -1142,6 +1142,12 @@
   </data>
   <data name="btn_TripSwitchOnOff.TabIndex" type="System.Int32, mscorlib">
     <value>8</value>
+  </data>
+  <data name="btn_TripSwitchOnOff.Text" xml:space="preserve">
+    <value>Trip on off</value>
+  </data>
+  <data name="btn_TripSwitchOnOff.TextAlign" type="System.Drawing.ContentAlignment, System.Drawing">
+    <value>BottomCenter</value>
   </data>
   <data name="&gt;&gt;btn_TripSwitchOnOff.Name" xml:space="preserve">
     <value>btn_TripSwitchOnOff</value>
